Removed line length from CPPLINT.cfg
With this edit, the maximum line length will be 80 characters instead of the earlier 240 characters we specified. Before we ship the next version of Lotus we need to check if we should add, remove, or edit anything. As our standards for our code have changed drastically since the first versions we should maybe check if we should keep those filter settings we currently have for example.